DOOM under MS-DOS 6.22, through EMM386's VCPI and over plain HIMEM
DOS/4GW now starts DOOM on the booted machine both ways: through VCPI
under EMM386 (its own page tables, flat 32-bit PM translated by the JIT)
and switching itself over XMS. -V clean through ST_Init under EMM386;
tests/boot/run.sh covers both.

- pc_bios: the master 8259 gets its registers. ICW1-4 set the vector
  base (IRQs deliver at base + n), OCW3 selects what port 20h reads (IRR,
  or the ISR after 0Bh), and the mask is honoured. DOS/4GW tells IRQ 0
  from a double fault, both vector 8, by reading the ISR; port 20h read 0.
- dbt_a64: a paged code page mapped below its linear address (DOOM at
  498000 on physical 3ED000) was decoded from mem + lin + delta with the
  32-bit delta added to the pointer unwrapped: 4 GB past guest memory,
  garbage blocks. The sum wraps in 32 bits now. pgtest T0E runs code on
  such a page (QEMU agrees; the old JIT took #UD).
- x86_deliver_exception calls the exception trace hook, so X86_EXC_TRACE
  sees faults raised in translated code; the trace prints CR2 for #PF.
- x86_exec_decoded clears pg_super after a helper, as x86_step does.

+++ dbt/dbt_a64.c
@@ -2839,7 +2839,7 @@ uint8_t *dbt_translate_block(x86_dbt *dbt, uint64_t key) {
     s_esnull = (key & KEY_ESNULL) != 0;
     s_dsnull = (key & KEY_DSNULL) != 0;
     s_pg_user = s_flat && s_paged && (cpu->seg[S_CS].sel & 3) == 3;
-    uint32_t code_page = 0, code_delta = 0;
+    uint32_t code_page = 0, code_delta = 0;     /* physical - linear, mod 2^32: add it before indexing mem */
     if (s_paged) {
         /* A paged block (V86, or flat protected mode) stays on one code
          * page. Its key is linear; its bytes, and their code-bitmap
@@ -2882,7 +2882,7 @@ uint8_t *dbt_translate_block(x86_dbt *dbt, uint64_t key) {
         if (s_flat && s_paged) {
             uint32_t lin = cpu->seg[S_CS].base + ip;
             if ((lin & 0xFFFFF000u) != code_page) break;
-            memcpy(buf, cpu->mem + lin + code_delta, 16);
+            memcpy(buf, cpu->mem + (uint32_t)(lin + code_delta), 16);
         } else if (s_flat) {
             /* flat CS: no limit to hit short of 4 GB; stay inside memory */
             uint32_t lin = cpu->seg[S_CS].base + ip;
@@ -2891,7 +2891,7 @@ uint8_t *dbt_translate_block(x86_dbt *dbt, uint64_t key) {
         } else if (s_paged) {
             uint32_t lin = cpu->seg[S_CS].base + ip;
             if ((lin & 0xFFFFF000u) != code_page) break;
-            memcpy(buf, cpu->mem + lin + code_delta, 16);          /* identity: the mirror applies A20 */
+            memcpy(buf, cpu->mem + (uint32_t)(lin + code_delta), 16);          /* identity: the mirror applies A20 */
         } else {
             fetch_at(cpu, ip, buf);
         }
